Make more generic.
This was cloned from a repo where they switched
the terms for the quadrants and their locations.
This switches them back to the same as the original.

I.e. you can switch the terms now so that the
presentation is different, but the underlying code
treats it all the same.